Change linkerd proxy default log format
This is a small change which sets the linkerd proxy to by default output
logs as json instead of plain text. The reasoning is that log collectors
such as datadog prefer structured log data so the majority of users
prefer this behavior instead of plain text. It is still possible to
change the format to plain text with an annotation on the pod.