azure/core: Fix bug in route table association (#532)
* azure/core: Fix bug in route table association

This does not affect XKF by default, only if custom routes are
configured.

We were looping through var.routes which contains a list of all the
route+subnet combinations, which would work for the first route but not
for the second and third because the subnet already had a route table
associated.

With this fix, we are moving the configuration to use the correct
var.route_config instead which is the same as the route_tabe
configuration.

This shouldn't cause any big issue, but will temporarily remove the
route association from the subnet and then add it again. But after this
fix, we will be able to use multiple routes for a route table - as god
intended it to be. :^)

* Add [Breaking]  to changelog

This is a standard that we have decided on internally. It should probably be documented as well :P

Co-authored-by: Edvin N <edvin.norling@xenit.se>