Exclude namespaces from Datadog (#648)
Prior to this change, when enabled, Datadog is configured to ship logs
from all tenant namespace plus a configured number of "extra" namespaces
through the "extra_namespaces" variable in eks_core and aks_core.
Now we also must be able to exclude namespaces from the tenant
namespaces. Instead of adding a "exclude_namespaces" variable, the
module interface is changed by removing the "extra_namespaces" variable
and an new "namespaces" variable is introduced where you configure all
namespaces that Datadog shall shipå logs from explicitly. In this way
the module interface will be easier to understand compared to the other
alternative. Also what namespace that are being used will be more
explicit though the configuration of the module.

Co-authored-by: Edvin N <edvin.norling@xenit.se>